backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OCR and image processing
pytesseract==0.3.10
Pillow==10.2.0
opencv-python-headless>=4.11.0
numpy<2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
rapidfuzz==3.6.1
